+++ modules/mp_downloader/mp_downloader/flasher.py
@@ -1,16 +1,7 @@
-import os
-import shutil
-import subprocess
-import time
-import sys
-
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple, Union
+from typing import  List, Optional, Tuple
 
-import bincopy
-import esptool
 import jsonlines
-import psutil
 import rich_click as click
 from loguru import logger as log
 from rich import print
@@ -18,11 +9,13 @@ from rich.table import Table
 
 from stubber.bulk.mpremoteboard import MPRemoteBoard
 
-from .common import DEFAULT_FW_PATH, PORT_FWTYPES
+from .common import DEFAULT_FW_PATH, FWInfo
+
+from .flash_uf2 import flash_uf2
+from .flash_esp import flash_esp
+from .flash_stm32 import flash_stm32
 
 # #########################################################################################################
-FWInfo = Dict[str, Union[str, bool]]
-
 
 def load_firmwares(fw_folder: Path) -> List[FWInfo]:
     """Load a list of available  firmwares from the jsonl file"""
@@ -93,317 +86,8 @@ def find_firmware(
     return fw_list
 
 
-# #########################################################################################################
-# Flash SAMD and RP2 via UF2
-# #########################################################################################################
-class UF2Disk:
-    """Info to support mounting and unmounting of UF2 drives on linux"""
-    device_path: str
-    label: str
-    mountpoint: str
-
-    def __repr__(self):
-        return repr(self.__dict__)
-
-def get_uf2_drives():
-    """
-    Get a list of all the (un)mounted UF2 drives
-    """
-    if sys.platform != "linux":
-        log.error("pumount only works on Linux")
-        return
-    # import blkinfo only on linux
-    from blkinfo import BlkDiskInfo
-
-    myblkd = BlkDiskInfo()
-    filters = {
-    'tran': 'usb',
-    }
-    usb_disks = myblkd.get_disks(filters)
-    for disk in usb_disks:
-        if disk["fstype"] == "vfat":
-            uf2_part = disk
-            # unpartioned usb disk or partition (e.g. /dev/sdb )
-            # SEEED WIO Terminal is unpartioned
-            # print( json.dumps(uf2_part, indent=4))
-            uf2 = UF2Disk()
-            uf2.device_path = "/dev/" + uf2_part["name"]
-            uf2.label = uf2_part["label"]
-            uf2.mountpoint = uf2_part["mountpoint"]
-            yield uf2 
-        elif disk["type"] == "disk" and disk.get("children") and len(disk.get("children")) > 0:
-            if disk.get("children")[0]["type"] == "part" and disk.get("children")[0]["fstype"] == "vfat":
-                uf2_part = disk.get("children")[0]
-                # print( json.dumps(uf2_part, indent=4))
-                uf2 = UF2Disk()
-                uf2.device_path = "/dev/" + uf2_part["name"]
-                uf2.label = uf2_part["label"]
-                uf2.mountpoint = uf2_part["mountpoint"]
-                yield uf2
-
-dismount_me : List[UF2Disk] = []
-
-def pmount(disk: UF2Disk):
-    """
-    Mount a UF2 drive if there is no mountpoint yet.
-    """
-    global dismount_me
-    if not disk.mountpoint:
-        if not disk.label:
-            disk.label = "UF2BOOT"
-        disk.mountpoint = f"/media/{disk.label}"
-        subprocess.run(["pmount",disk.device_path,  disk.mountpoint ])
-        log.info(f"Mounted {disk.label} at {disk.mountpoint}")
-        dismount_me.append(disk)
-    else:
-        log.warning(f"{disk.label} already mounted at {disk.mountpoint}")
-
-def pumount(disk: UF2Disk):
-    """
-    Unmount a UF2 drive
-    """
-    if sys.platform != "linux":
-        log.error("pumount only works on Linux")
-        return
-    if disk.mountpoint:
-        subprocess.run(["pumount", disk.mountpoint]) # ), f"/media/{disk.label}"])
-        log.info(f"Unmounted {disk.label} from {disk.mountpoint}")
-        disk.mountpoint = f""
-    else:
-        log.warning(f"{disk.label} already dismounted")
-
-def dismount_uf2():
-    global dismount_me
-    for disk in dismount_me:
-        pumount(disk)
-    dismount_me = []
-
-def flash_uf2(mcu: MPRemoteBoard, fw_file: Path) -> Optional[MPRemoteBoard]:
-    """
-    Flash .UF2 devices via bootloader and filecopy
-    - mpremote bootloader
-    - Wait for the device to mount as a drive (up to 5s)
-    - detect new drive with INFO_UF2.TXT
-    - copy the firmware file to the drive
-    - wait for the device to restart (5s)
-
-    for Lunix : 
-    pmount and pumount are used to mount and unmount the drive
-    as this is not done automatically by the OS in headless mode.
-    """
-    if PORT_FWTYPES[mcu.port] not in [".uf2"]:
-        log.error(f"UF2 not supported on {mcu.board} on {mcu.serialport}")
-        return None
-
-    log.info(f"Entering UF2 bootloader on {mcu.board} on {mcu.serialport}")
-    mcu.run_command("bootloader", timeout=10)
-
-    if sys.platform == "linux":
-        destination = wait_for_UF2_linux()
-    elif sys.platform == "win32":
-        destination = wait_for_UF2_windows()
-    else:
-        log.error(f"OS {sys.platform} not supported")
-        return None
-
-    if not destination or not destination.exists() or not (destination / "INFO_UF2.TXT").exists():
-        log.error("Board is not in bootloader mode")
-        return None
-
-    log.info("Board is in bootloader mode")
-    log.info(f"Copying {fw_file} to {destination}")
-    shutil.copy(fw_file, destination)
-    log.success("Done copying, resetting the board and wait for it to restart")
-    if sys.platform == "linux":
-        dismount_uf2()
-    time.sleep(5 * 2) # 5 secs to short on linux
-    return mcu
-
-def wait_for_UF2_linux():
-    destination = ""
-    wait = 10
-    uf2_drives = []
-    while not destination and wait > 0:
-        log.info(f"Waiting for mcu to mount as a drive : {wait} seconds left")
-        uf2_drives += list(get_uf2_drives())
-        for drive in get_uf2_drives():
-            pmount(drive)
-            time.sleep(1)
-            if Path(drive.mountpoint, "INFO_UF2.TXT").exists():
-                # board_id = get_board_id(Path(drive.mountpoint))
-                destination = Path(drive.mountpoint)
-                break
-        time.sleep(1)
-        wait -= 1
-    return destination
-
-def wait_for_UF2_windows():
-    destination = ""
-    wait = 10
-    while not destination and wait > 0:
-        log.info(f"Waiting for mcu to mount as a drive : {wait} seconds left")
-        drives = [drive.device for drive in psutil.disk_partitions()]
-        for drive in drives:
-            if Path(drive, "INFO_UF2.TXT").exists():
-                board_id = get_board_id(Path(drive))
-                destination = Path(drive)
-                break
-        time.sleep(1)
-        wait -= 1
-    return destination
-
-def get_board_id(path:Path):
-    # Option : read Board-ID from INFO_UF2.TXT
-    board_id = "Unknown"
-    with open(path /  "INFO_UF2.TXT") as f:
-        data = f.readlines()
-    for line in data:
-        if line.startswith("Board-ID"):
-            board_id = line[9:].strip()
-    log.trace(f"Found Board-ID={board_id}")
-    return board_id
 
 
-# #########################################################################################################
-# Flash ESP32 and ESP8266 via esptool
-# #########################################################################################################
-
-
-def flash_esp(mcu: MPRemoteBoard, fw_file: Path, *, erase_flash: bool = True) -> Optional[MPRemoteBoard]:
-    if mcu.port not in ["esp32", "esp8266"] or mcu.board in ["ARDUINO_NANO_ESP32"]:
-        log.error(f"esptool not supported for {mcu.port} {mcu.board} on {mcu.serialport}")
-        return None
-
-    log.info(f"Flashing {fw_file} on {mcu.board} on {mcu.serialport}")
-    if mcu.port == "esp8266":
-        baud_rate = str(460_800)
-    else:
-        baud_rate = str(512_000)
-        # baud_rate = str(115_200)
-    cmds: List[List[str]] = []
-    if erase_flash:
-        cmds.append(f"esptool --chip {mcu.cpu} --port {mcu.serialport} erase_flash".split())
-
-    if mcu.cpu.upper() in ("ESP32", "ESP32S2"):
-        start_addr = "0x1000"
-    elif mcu.cpu.upper() in ("ESP32S3", "ESP32C3"):
-        start_addr = "0x0"
-    if mcu.cpu.upper().startswith("ESP32"):
-        cmds.append(
-            f"esptool --chip {mcu.cpu} --port {mcu.serialport} -b {baud_rate} write_flash --compress {start_addr}".split()
-            + [str(fw_file)]
-        )
-    elif mcu.cpu.upper() == "ESP8266":
-        start_addr = "0x0"
-        cmds.append(
-            f"esptool --chip {mcu.cpu} --port {mcu.serialport} -b {baud_rate} write_flash --flash_size=detect {start_addr}".split()
-            + [str(fw_file)]
-        )
-    try:
-        for cmd in cmds:
-            log.info(f"Running {' '.join(cmd)} ")
-            esptool.main(cmd[1:])
-    except Exception as e:
-        log.error(f"Failed to flash {mcu.board} on {mcu.serialport} : {e}")
-        return None
-
-    log.info("Done flashing, resetting the board and wait for it to restart")
-    time.sleep(5)
-    mcu.get_mcu_info()
-    log.success(f"Flashed {mcu.version} to {mcu.board} on {mcu.serialport} done")
-    return mcu
-
-
-# #########################################################################################################
-# flash STM32 using STM32CubeProgrammer
-# needs to be installed independenty from https://www.st.com/en/development-tools/stm32cubeprog.html
-# #########################################################################################################
-
-
-def get_stm32_start_address(fw_file: Path):
-    """
-    Get the start address of the firmware file, to allow automatic restart from that address after flashing
-    """
-    try:
-        fw_hex = bincopy.BinFile(str(fw_file))
-        return f"0x{fw_hex.execution_start_address:08X}"
-    except Exception:
-
-        return ""
-
-
-def flash_stm32(mcu: MPRemoteBoard, fw_file: Path, *, erase_flash: bool = True) -> Optional[MPRemoteBoard]:
-    """
-    Flash STM32 devices using STM32CubeProgrammer CLI
-    - Enter bootloader mode
-    - wait 2s for the device to be detected
-    - list the connected DFU devices
-    """
-    STM32_CLI = "C:\\Program Files\\STMicroelectronics\\STM32Cube\\STM32CubeProgrammer\\bin\\STM32_Programmer_CLI.exe"
-
-    if not Path(STM32_CLI).exists():
-        log.error(
-            f"STM32CubeProgrammer not found at {STM32_CLI}\nPlease install it from https://www.st.com/en/development-tools/stm32cubeprog.html"
-        )
-        return None
-
-    log.info(f"Entering STM bootloader on {mcu.board} on {mcu.serialport}")
-    mcu.run_command("bootloader")
-    time.sleep(2)
-    # run STM32_Programmer_CLI.exe --list
-    cmd = [
-        STM32_CLI,
-        "--list",
-    ]
-    results = subprocess.run(cmd, capture_output=True, text=True).stdout.splitlines()
-    if "Product ID             : STM32  BOOTLOADER" in results:
-        log.error("No STM32 BOOTLOADER detected")
-        return None
-    echo = False
-    for line in results:
-        if line.startswith("=====  DFU Interface"):
-            echo = True
-        if line.startswith("===== STLink"):
-            echo = False
-        if echo:
-            print(line)
-    # Try to connect - no action
-    cmd = [
-        STM32_CLI,
-        "--connect",
-        "port=USB1",
-    ]
-    results = subprocess.run(cmd, capture_output=True, text=True).stdout.splitlines()
-    if erase_flash:
-        log.info("Erasing flash")
-        cmd = [
-            STM32_CLI,
-            "--connect",
-            "port=USB1",
-            "--erase",
-            "all",
-        ]
-        results = subprocess.run(cmd, capture_output=True, text=True).stdout.splitlines()
-
-    log.info("Flashing")
-    start_address = get_stm32_start_address(fw_file)
-
-    log.trace(f"STM32_Programmer_CLI.exe --connect port=USB1 --write {str(fw_file)} --go {start_address}")
-    # !"C:\Program Files\STMicroelectronics\STM32Cube\STM32CubeProgrammer\bin\STM32_Programmer_CLI.exe" --connect port=USB1 --write "{str(fw_file)}" --go {start_address}
-    cmd = [
-        STM32_CLI,
-        "--connect",
-        "port=USB1",
-        "--write",
-        str(fw_file),
-        "--go",
-        start_address,
-    ]
-    results = subprocess.run(cmd, capture_output=True, text=True).stdout.splitlines()
-    log.success("Done flashing, resetting the board and wait for it to restart")
-    time.sleep(5)
-    mcu.get_mcu_info()
-    return mcu
 
 
 # #########################################################################################################
