+++ modules/mp_downloader/mp_downloader/uf2_boardid.py
@@ -1,0 +1,13 @@
+from pathlib import Path
+from loguru import logger as log
+
+def get_board_id(path:Path):
+    # Option : read Board-ID from INFO_UF2.TXT
+    board_id = "Unknown"
+    with open(path /  "INFO_UF2.TXT") as f:
+        data = f.readlines()
+    for line in data:
+        if line.startswith("Board-ID"):
+            board_id = line[9:].strip()
+    log.trace(f"Found Board-ID={board_id}")
+    return board_id
